Move views into their own namespace
