+++ lib/mongo/operation/write/ensure_index.rb
@@ -80,7 +80,7 @@ module Mongo
 
         def message(index)
           index_spec = options.merge(ns: namespace, key: index, name: index_name)
-          Protocol::Insert.new(db_name, IndexView::COLLECTION, [ index_spec ])
+          Protocol::Insert.new(db_name, View::Index::COLLECTION, [ index_spec ])
         end
       end
     end
