+++ lib/mongo/view.rb
@@ -12,26 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'forwardable'
-require 'bson'
-require 'mongo/loggable'
-require 'mongo/logger'
-require 'mongo/errors'
-require 'mongo/event'
-require 'mongo/auth'
-require 'mongo/client'
-require 'mongo/cluster'
-require 'mongo/collection'
-require 'mongo/connection'
-require 'mongo/cursor'
-require 'mongo/database'
-require 'mongo/grid'
-require 'mongo/operation'
-require 'mongo/pool'
-require 'mongo/protocol'
-require 'mongo/server'
-require 'mongo/server_preference'
-require 'mongo/socket'
-require 'mongo/uri'
-require 'mongo/version'
-require 'mongo/view'
+require 'mongo/view/executable'
+require 'mongo/view/collection'
+require 'mongo/view/index'
