+++ spec/support/shared/cursor.rb
@@ -18,7 +18,7 @@ shared_context 'shared cursor' do
   end
 
   let(:view_opts) { {} }
-  let(:view) { Mongo::CollectionView.new(collection, {}, view_opts) }
+  let(:view) { Mongo::View::Collection.new(collection, {}, view_opts) }
 
   let(:nonzero) { 1 }
   let(:b) { proc { |d| d } }
